+++ WeatherAPI/src/main/java/neilsayok/github/weatherapi/data/Daily.kt
@@ -9,7 +9,7 @@ data class Daily(
     @SerializedName("dew_point")
     var dewPoint: Double? = 0.0,
     @SerializedName("dt")
-    var dt: Int? = 0,
+    var dt: Long? = 0,
     @SerializedName("feels_like")
     var feelsLike: FeelsLike? = FeelsLike(),
     @SerializedName("humidity")
