+++ WeatherAPI/src/main/java/neilsayok/github/weatherapi/data/Minutely.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Minutely(
     @SerializedName("dt")
-    var dt: Int? = 0,
+    var dt: Long? = 0,
     @SerializedName("precipitation")
     var precipitation: Int? = 0
 )
